+++ packages/calcite-components/src/components/notice/notice.tsx
@@ -12,13 +12,13 @@ import {
   stringOrBoolean,
 } from "@arcgis/lumina";
 import { setRequestedIcon, slotChangeHasAssignedElement } from "../../utils/dom";
-import { componentFocusable } from "../../utils/component";
 import { Kind, Scale, Width } from "../interfaces";
 import { KindIcons } from "../resources";
 import { toggleOpenClose, OpenCloseComponent } from "../../utils/openCloseComponent";
 import { getIconScale } from "../../utils/component";
 import { IconNameOrString } from "../icon/interfaces";
 import { useT9n } from "../../controllers/useT9n";
+import { useSetFocus } from "../../controllers/useSetFocus";
 import T9nStrings from "./assets/t9n/messages.en.json";
 import { CSS, SLOTS } from "./resources";
 import { styles } from "./notice.scss";
@@ -66,6 +66,8 @@ export class Notice extends LitElement implements OpenCloseComponent {
    */
   messages = useT9n<typeof T9nStrings>();
 
+  private focusSetter = useSetFocus<this>()(this);
+
   //#endregion
 
   //#region State Properties
@@ -110,18 +112,10 @@ export class Notice extends LitElement implements OpenCloseComponent {
   /** Sets focus on the component's first focusable element. */
   @method()
   async setFocus(): Promise<void> {
-    await componentFocusable(this);
-
-    const noticeLinkEl = this.el.querySelector("calcite-link");
-
-    if (!this.closeButton.value && !noticeLinkEl) {
-      return;
-    }
-    if (noticeLinkEl) {
-      return noticeLinkEl.setFocus();
-    } else if (this.closeButton.value) {
-      this.closeButton.value.focus();
-    }
+    return this.focusSetter(() => {
+      const noticeLinkEl = this.el.querySelector("calcite-link");
+      return noticeLinkEl || this.closeButton.value;
+    });
   }
 
   //#endregion