+++ packages/calcite-components/src/components/radio-button/radio-button.tsx
@@ -2,7 +2,7 @@
 import { PropertyValues } from "lit";
 import { LitElement, property, createEvent, h, method, JsxNode } from "@arcgis/lumina";
 import { getRoundRobinIndex } from "../../utils/array";
-import { focusElement, getElementDir } from "../../utils/dom";
+import { getElementDir } from "../../utils/dom";
 import {
   CheckableFormComponent,
   connectForm,
@@ -15,9 +15,9 @@ import {
   updateHostInteraction,
 } from "../../utils/interactive";
 import { connectLabel, disconnectLabel, getLabelText, LabelableComponent } from "../../utils/label";
-import { componentFocusable } from "../../utils/component";
 import { Scale } from "../interfaces";
 import type { Label } from "../label/label";
+import { useSetFocus } from "../../controllers/useSetFocus";
 import { CSS } from "./resources";
 import { styles } from "./radio-button.scss";
 
@@ -50,6 +50,8 @@ export class RadioButton
   labelEl: Label["el"];
 
   private rootNode: HTMLElement;
+
+  private focusSetter = useSetFocus<this>()(this);
 
   // #endregion
 
@@ -125,11 +127,9 @@ export class RadioButton
   /** Sets focus on the component. */
   @method()
   async setFocus(): Promise<void> {
-    await componentFocusable(this);
-
-    if (!this.disabled) {
-      focusElement(this.containerEl);
-    }
+    return this.focusSetter(() => {
+      return this.containerEl;
+    });
   }
 
   // #endregion
@@ -412,8 +412,6 @@ export class RadioButton
     ).filter((radioButton) => radioButton.name === this.name);
     let currentIndex = 0;
 
-    const radioButtonsLength = radioButtons.length;
-
     radioButtons.some((item, index) => {
       if (item.checked) {
         currentIndex = index;
@@ -427,17 +425,35 @@ export class RadioButton
         event.preventDefault();
         this.selectItem(
           radioButtons,
-          getRoundRobinIndex(Math.max(currentIndex - 1, -1), radioButtonsLength),
+          this.getNextNonDisabledIndex(radioButtons, currentIndex, "left"),
         );
         return;
       case "ArrowRight":
       case "ArrowDown":
         event.preventDefault();
-        this.selectItem(radioButtons, getRoundRobinIndex(currentIndex + 1, radioButtonsLength));
+        this.selectItem(
+          radioButtons,
+          this.getNextNonDisabledIndex(radioButtons, currentIndex, "right"),
+        );
         return;
       default:
         return;
     }
+  }
+
+  private getNextNonDisabledIndex(
+    radioButtons: RadioButton["el"][],
+    startIndex: number,
+    dir: "left" | "right",
+  ): number {
+    const totalButtons = radioButtons.length;
+    const offset = dir === "left" ? -1 : 1;
+    let selectIndex = getRoundRobinIndex(startIndex + offset, totalButtons);
+    while (radioButtons[selectIndex].disabled) {
+      selectIndex = getRoundRobinIndex(selectIndex + offset, totalButtons);
+    }
+
+    return selectIndex;
   }
 
   private onContainerBlur(): void {