+++ packages/calcite-components/src/components/table-row/table-row.tsx
@@ -274,11 +274,11 @@ export class TableRow extends LitElement implements InteractiveComponent {
           event.preventDefault();
           break;
         case "ArrowLeft":
-          focusElementInGroup(cells, el, "previous", false);
+          focusElementInGroup(cells, el, "previous", false, false);
           event.preventDefault();
           break;
         case "ArrowRight":
-          focusElementInGroup(cells, el, "next", false);
+          focusElementInGroup(cells, el, "next", false, false);
           event.preventDefault();
           break;
         case "Home":
@@ -286,7 +286,7 @@ export class TableRow extends LitElement implements InteractiveComponent {
             this.emitTableRowFocusRequest(1, this.positionAll, "first");
             event.preventDefault();
           } else {
-            focusElementInGroup(cells, el, "first", false);
+            focusElementInGroup(cells, el, "first", false, false);
             event.preventDefault();
           }
           break;
@@ -295,7 +295,7 @@ export class TableRow extends LitElement implements InteractiveComponent {
             this.emitTableRowFocusRequest(this.rowCells?.length, this.positionAll, "last", true);
             event.preventDefault();
           } else {
-            focusElementInGroup(cells, el, "last", false);
+            focusElementInGroup(cells, el, "last", false, false);
             event.preventDefault();
           }
           break;