+++ packages/calcite-tailwind-preset/lint-staged.config.mjs
@@ -1,6 +1,14 @@
+import { resolve } from "node:path";
 import baseConfig from "../../lint-staged.config.mjs";
 
 export default {
   ...baseConfig,
-  "*.{ts,tsx,mjs,cjs}": ["eslint --fix", "prettier --write"],
+  "*.{ts,tsx,mjs,cjs}": (absolutePaths) => {
+    const files = absolutePaths.join(" ");
+    const { dirname } = import.meta;
+    return [
+      `eslint --fix ${files}`,
+      `prettier --ignore-path ${resolve(dirname, "../../.prettierignore")} --write ${files}`,
+    ];
+  },
 };